refactor: centralize section schemas
